+++ src/ludamus/links/db/django/uow.py
@@ -5,7 +5,7 @@ from django.contrib.auth import login as django_login
 from django.db import transaction
 
 from ludamus.adapters.db.django.models import User
-from ludamus.links.db.django import repositories
+from ludamus.links.db.django import crowd, repositories
 from ludamus.links.db.django.agenda_item import AgendaItemRepository
 from ludamus.links.db.django.schedule_change_log import ScheduleChangeLogRepository
 from ludamus.pacts import UnitOfWorkProtocol, UserType
@@ -27,20 +27,20 @@ class UnitOfWork(UnitOfWorkProtocol):  # noqa: PLR0904
         django_login(request, user)
 
     @cached_property
-    def active_users(self) -> repositories.UserRepository:
-        return repositories.UserRepository(user_type=UserType.ACTIVE)
+    def active_users(self) -> crowd.UserRepository:
+        return crowd.UserRepository(user_type=UserType.ACTIVE)
 
     @cached_property
     def agenda_items(self) -> AgendaItemRepository:
         return AgendaItemRepository()
 
     @cached_property
-    def anonymous_users(self) -> repositories.UserRepository:
-        return repositories.UserRepository(user_type=UserType.ANONYMOUS)
+    def anonymous_users(self) -> crowd.UserRepository:
+        return crowd.UserRepository(user_type=UserType.ANONYMOUS)
 
     @cached_property
-    def connected_users(self) -> repositories.ConnectedUserRepository:
-        return repositories.ConnectedUserRepository()
+    def connected_users(self) -> crowd.ConnectedUserRepository:
+        return crowd.ConnectedUserRepository()
 
     @cached_property
     def event_proposal_settings(self) -> repositories.EventProposalSettingsRepository:
